% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_extraction.R
\name{codon_extraction}
\alias{codon_extraction}
\title{Codon extraction}
\usage{
codon_extraction(dna_sample, start = 1)
}
\arguments{
\item{dna_sample}{DNA sample used for extracting codons}

\item{start}{Starting position for extracting codons}
}
\value{
Vector named codons that contains the extracted codons from the DNA sample
}
\description{
Codon extraction
}
\examples{
function that extracts codons from a DNA sample
}
