% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_iden.R
\name{codon_iden}
\alias{codon_iden}
\title{Codon identifier}
\format{
Function
}
\usage{
codon_iden(RNA_seq, start = 1)
}
\arguments{
\item{RNA_seq}{}

\item{start}{}
}
\value{

}
\description{
Identifies codons from inputted RNA sequence.
}
