% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_table.R
\docType{data}
\name{codon_table}
\alias{codon_table}
\title{Standard Codon table}
\format{
data frame
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
codon_table
}
\description{
Translation table
}
\keyword{datasets}
