% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcatenateCodons.R
\name{ConcatenateCodons}
\alias{ConcatenateCodons}
<<<<<<< HEAD
\title{Concatenate amino acids values of given key codons into one string}
\usage{
ConcatenateCodons(key_codons)
}
\arguments{
\item{key_codons}{list of key codons from which the amino acids will be merged}
}
\value{
string of amino acids concatenated
}
\description{
Concatenate amino acids values of given key codons into one string
}
\examples{
none
=======
\title{Concatenate codons into one string}
\usage{
ConcatenateCodons(codons2)
}
\arguments{
\item{codons2}{list of codons to be merged}
}
\value{
string of codons concatenated
}
\description{
This function concatenates all the codons into a single string
}
\examples{
none

>>>>>>> a67b3e519b95d6dc389c7fedf50da980c02c2960
}
