% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_dna_seq.R
\name{create_random_dna_seq}
\alias{create_random_dna_seq}
\title{Create random DNA seq}
\usage{
create_random_dna_seq(seq_length)
}
\arguments{
\item{seq_length}{A positive integer indicating the
length of the desired DNA sequence.}
}
\value{
A random DNA sequence as string with length \code{seq_length}.
}
\description{
Lets the user create a string of a randomized DNA-sequence.
}
