% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_freq_plot.R
\name{get_freq_plot}
\alias{get_freq_plot}
\title{Generate frequency plot}
\usage{
get_freq_plot(aa_seq)
}
\arguments{
\item{aa_seq}{An Amino acid sequence as string.}
}
\value{
A frequency plot
}
\description{
Takes amino acid sequence as input, counts instances of amino acids and
generates a frequency plot.
}
