% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleDNA.R
\name{SampleDNA}
\alias{SampleDNA}
\title{DNA chain generator}
\usage{
SampleDNA(size)
}
\arguments{
\item{size}{size of the desired DNA chain.}
}
\value{
It will give a random DNA chain with the desired size.
}
\description{
DNA chain generator
}
\examples{
40
}
