% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptionDNA.R
\name{TranscriptionDNA}
\alias{TranscriptionDNA}
\title{DNA Transcription}
\usage{
TranscriptionDNA(DNA)
}
\arguments{
\item{DNA}{string of DNA to be transcribed}
}
\value{
returns string object with RNA transcribed from the input DNA.
}
\description{
DNA Transcription
}
