% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_codon.R
\name{translate_codon}
\alias{translate_codon}
\title{Translate codons to Amino Acis}
\usage{
translate_codon(codons)
}
\arguments{
\item{codons}{A vector with codons}
}
\value{
A string with amino acids
}
\description{
Takes a vector with codons, and translate to amino acids according to \code{codon_table}
}
