% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcatenateCodons.R
\name{ConcatenateCodons}
\alias{ConcatenateCodons}
\title{Concatenate amino acids values of given key codons into one string}
\usage{
ConcatenateCodons(key_codons)
}
\arguments{
\item{key_codons}{list of key codons from which the amino acids will be merged}
}
\value{
string of amino acids concatenated
}
\description{
Concatenate amino acids values of given key codons into one string
}
\examples{
none
}
