% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConcatenateCodons.R
\name{ConcatenateCodons}
\alias{ConcatenateCodons}
\title{Concatenate codons into one string}
\usage{
ConcatenateCodons(codons2)
}
\arguments{
\item{codons2}{list of codons to be merged}
}
\value{
string of codons concatenated
}
\description{
This function concatenates all the codons into a single string
}
\examples{
none

}
